refactor(pytorch): extract prefix cache sequence state
