+++ lmdeploy/pytorch/paging/block_trie/trie.py
@@ -73,7 +73,8 @@ from dataclasses import dataclass
 
 import numpy as np
 
-from lmdeploy.pytorch.messages import PrefixCacheExtraHashes, SchedulerSequence
+from lmdeploy.pytorch.messages import SchedulerSequence
+from lmdeploy.pytorch.prefix_cache_state import PrefixCacheExtraHashes
 from lmdeploy.utils import get_logger
 
 from ...config import CacheConfig
